Moved GameConfig and PlayerConfig out of ui package
- applied refactor operation to update all files where PlayerConfig and
GameConfig are used with the appropriate package name
