+++ src/main/java/net/demilich/metastone/game/Player.java
@@ -14,7 +14,7 @@ import net.demilich.metastone.game.entities.heroes.Hero;
 import net.demilich.metastone.game.entities.minions.Minion;
 import net.demilich.metastone.game.logic.CustomCloneable;
 import net.demilich.metastone.game.statistics.GameStatistics;
-import net.demilich.metastone.gui.gameconfig.PlayerConfig;
+import net.demilich.metastone.game.gameconfig.PlayerConfig;
 
 public class Player extends CustomCloneable {
