+++ src/main/java/net/demilich/metastone/gui/battleofdecks/StartBattleOfDecksCommand.java
@@ -23,7 +23,7 @@ import net.demilich.metastone.game.cards.CardSet;
 import net.demilich.metastone.game.decks.Deck;
 import net.demilich.metastone.game.decks.DeckFormat;
 import net.demilich.metastone.game.logic.GameLogic;
-import net.demilich.metastone.gui.gameconfig.PlayerConfig;
+import net.demilich.metastone.game.gameconfig.PlayerConfig;
 
 public class StartBattleOfDecksCommand extends SimpleCommand<GameNotification> {
